begin hooking igs_m027_023vid.cpp to PGM video system
